pyproject.toml
[project]

name = "movie2sub"
version = "0.0.1"
authors = [
    { name="Adrian Mincu", email="mincu.adrian13@gmail.com" },
    { name="Ciprian Hutanu", email="eusebiuciprian.hutanu03@gmail.com" }
]

description = "Movie2Sub is a Python-based package used to generate movie subtitles from an input audio."
readme = "README.md"

requires-python = ">=3.10,<3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

dependencies = [
    'python-dotenv~=1.1',
    'qbittorrent-api~=2025.5',
    'beautifulsoup4~=4.13',
    'fake-useragent~=2.2',
    'pysrt~=1.1',
    'numpy<2.0',
    'scipy~=1.15',
    'soundfile~=0.13',
    'tqdm~=4.67',
    'spleeter~=2.4',
]

[project.optional-dependencies]
dev = [
    'pytest~=8.3',
    'black~=24.10',
    'isort~=5.13',
]

[tool.pytest.ini_options]
minversion = "8.3"
addopts = "-ra"
testpaths = [
    "tests",
    "integration",
]

[project.urls]
Homepage = "https://github.com/skpha13/stt-movie-subs"
Issues = "https://github.com/skpha13/stt-movie-subs/issues"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["movie2sub", "movie2sub.*"]